Worked more on the home page
I made the middle of the page fit properly and formatted everything in
the middle of the to make it look how it is on the wireframe. I also
fixed a bit of the nav bar to make the nav buttons properly spaced
between each other, however, that's as far as I got with that. WIll work
on it a bit later.
I also cropped the packet-s.jpg image for easier manipulation on the
webpage and added a favicon because I thought that it would be cool to
have one.