Minor fixups
I finished the header/nav bar and implimented it in the signup page.
Added a footer just in case for next time when Johsnon tells me whether
I need to have a contact us/privacy policy/terms and conditions etc.
that's where I'll put it all.
I cropped the small packet of water again to make it more centre.