Database stuff
I made 3 more tables as shown on the database wireframe thing and renamed the members table to users. This meant that I had to change my code in the login, signup and update pages. Also moved productdetails.xlsx from the root folder to the database fodler since it has info for the database. The dbTables.xlsx shows which tables and fields and stuff are in which table in the database.