Changed table rows/columns in database and minor text fixes
I deleted many things from the database such as productname and price in the cart table because I didn't fully understand how it worked. Apparently I didn't need it because it would take the data from the productID which is cool. Also changed "Jar" to "Flask" in shop.php because it's supposed to be a flask