Added comments, TODO and made nav.php
Added comments on (almost?) all pages.
Added a TODO list on every page that needs things to be done.
Added nav.php which is just the php code for the changing nav bar. This makes my code more efficient since it just calls on the nav.php instead of having the same code on the shop and index pages.