Added the cart
CSS:
I made the product card's buttons look good and fit with the theme.
shop.php:
I copy/pasted the cart, changed variables to make them closer to what I'd use and tinkered with the cart until it worked. I fiddled around for about 5 hours getting this damned thing to work and it finally is! Marvellous!
cart.php:
Added a seperate page for the cart. I want to use a seperate page since I think that it will look better and it's not hard to implement.
conn.php:
Added comments.
Next steps:
Make the cart look better, allow users to access the shop even if not logged in, only allow users to add to and visit the cart if logged in, delete the cart from shop.php, possibly add a "Welcome USER". I tried to do this, however, I was met with many many bugs, so I decided to postpone it. Perhaps indefinitely.