MAJOR CHANGES
CSS:
Made everything more pretty?
I was having trouble with the table's borders, they weren't going away, and googled the problem. 10 minutes later I find a solution: border-collapse. CSS sucks :smile:
Added back the cart button styling because I added the button back.
cart.php:
Made it look pretty. To a degree. A good degree. Still a degree. But a good degree :accept:
I pretty much just edited some text and made it so that the images of the items also showed.
index.php:
Removed almost everything in the header. (Minor text fixes.)
login.php:
Minor text fixes.
nav.php:
Moved all of the header into here so that it takes up less overall space. It is all in one document instead of being spread out among many. Still functions the same way, though, where nav.php is called and the header shows.
Also added a cart button that leads the user to the cart.
shop.php:
Added an IF statement that makes sure that the user is logged in before being able to add to cart. The user can also visit the shop before logging in, however, must first log in before adding to cart.
Made it so that the image was able to be seen in the cart.
Removed the cart that was at the bottom of the page.
signup.php:
Minor text fixes.
Update.php:
Minor text fixes.
TODO:
Add comments, finish styling cart.php, possibly add a WELCOME USER if I have time for it, TEST THE CRAP OUT OF IT yay... 🙃