FIXED THE PROBLEM!!!!!
I finally fixed the problem of the items going outside the cart thing! I FEEL GLORIOUS! I was about to give up and scrap it BUT YAY! What I did was move the closing bracket to above the end of the table where it should be and probably was originally. This fixed the issue. The problem was pretty much that the products were being placed outside the table.
Also fixed a minor bug where the first item added to cart didn't trigger the notification of "Item added to cart". I just added the same notification down below where the IF cart has nothing statement is.
GLORIOUS!