Added the WELCOME USER and commented a lot
Commented on the cart, shop and update pages.
Made things look prettier with CSS
Started the firstname session in the login page so that the WELCOME USER works and made the update page refresh after each update to make the WELCOME USER update because it doesn't update properly without it, it keep sthe old name until the session is started back up.