Did more testing for that EXCELLENCE
I did more testing and found that the characters ' and " could not be used since they interfere with the PHP that assigns the data to variables. This led me to finding the htmlspecialchars() function which I thought would be my saving grace and fix this problem by converting it into a normal character. I did not work. I think that this is because ' and " aren't special characters, they're normal characters that just like to screw up Dan's code. I didn't end up fixing this problem, however, commented on it in the testing document.